don't emit events to session listener directly
according to RepositoryConnector javadoc "a connector must not emit
any events to the transfer listener configured for the session. Instead,
transfer events must be emitted only to the listener (if any) specified
for a given download/upload request." Also confirmed with aether code
that session listener is set for individual tranfers, so the original
implementation delivered the same event twice.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>