Make sure Content-Length header is set of PUT requests
Our anonymous implementation of ResponseBody does not implement contentLength() so the content length
returned is always -1 which triggers the behaviour in OkHttp which sets the transfer-encoding to chunked.
This causes an IT in Maven to fail which looks to make sure the transfer-encoding is not set to chunked
because NGinx does not support chunked transfer-encoding without an extension module.

There's 3 ways to set the length. In descending order of preference they are content-length, chunked encoding, and EOF.
If you don't explicitly say chunked, the server will probably assume EOF. You really should specify either content-length
or chunked transfer-encoding.